+++ server/api/src/main/java/com/walmartlabs/concord/server/api/security/apikey/CreateApiKeyRequest.java
@@ -3,28 +3,35 @@ package com.walmartlabs.concord.server.api.security.apikey;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import javax.validation.constraints.NotNull;
 import java.io.Serializable;
 import java.util.UUID;
 
 public class CreateApiKeyRequest implements Serializable {
 
-    @NotNull
     private final UUID userId;
 
+    private final String username;
+
     @JsonCreator
-    public CreateApiKeyRequest(@JsonProperty("userId") UUID userId) {
+    public CreateApiKeyRequest(@JsonProperty("userId") UUID userId,
+                               @JsonProperty("username") String username) {
         this.userId = userId;
+        this.username = username;
     }
 
     public UUID getUserId() {
         return userId;
     }
 
+    public String getUsername() {
+        return username;
+    }
+
     @Override
     public String toString() {
         return "CreateApiKeyRequest{" +
                 "userId=" + userId +
+                ", username='" + username + '\'' +
                 '}';
     }
 }